Expose API via apache and reverse proxy from zeeguu-web
Currently the API is running in debug mode. This commit
ensures that the API will run as a WSGI module exposed
via an apache.

The Zeeguu-Web apache will act as a reverse proxy between
the clients and the API.

In development mode, the API will run in debug mode with
the python command.